Add legend textbox in queryctl editor
The legend textbox provides a mechanism for specifying a legend template
for each query in a panel. The template will support regular grafana
template variable substitutions as well as label substitutions using the
{{labelname}} syntax. If the legend textbox is empty, the query string
itself will be used as the template.

The remaining implementation for this needs changes in both server side
(pmproxy) and client side.