redis: pass correct timespec to pmproxy (add interval and drop samples argument)
Previously only the sample count was passed to pmproxy. If Grafana's
requested sample interval was different than the metric sample interval
logged by pmlogger, the graph was missing datapoints (or had too many
datapoints).

This changes passes the interval to pmproxy, and drops the sample count,
as it's not required (start + finish + interval should be sufficient).